+++ CHANGELOG.md
@@ -28,7 +28,30 @@ Released YYYY/MM/DD
 
 --------------------------------------------------------------------------------
 
-# 0.3.2
+# 0.3.5
+
+Released 2021/12/2
+
+## Changed
+
+* The LLVM mangling for vector types with a dimension expression is now
+  supported. The libiberty mangling which appears to be unused by gcc has
+  been removed.
+
+--------------------------------------------------------------------------------
+
+# 0.3.4
+
+Released 2021/11/21
+
+## Added
+
+* `DemangleOptions` now has a `hide_expression_literal_types` method that
+  can make it easier to match user-provided template instance names. #230
+
+--------------------------------------------------------------------------------
+
+# 0.3.3
 
 Released 2021/7/8